Made a change to fully explore the graph.
Previously, when exploring the edges of a node in the isPossible/DFS function, if we encountered a previously fully explored node, the function used to return. Doing this stops the exploration of other edges of the current node. This still worked as if the other edges contained sink nodes, there was no cycle and we wouldn't explore them anyhow in the DFS function, and the non sink nodes were still later explored by the CanFinish function. But if we want the complete topological ordering (LeetCode 210. Course Schedule II), returning here would result in the topological order for the current node in not being set and the other edges also not getting the correct order. For a proper implementation of DFS, we should continue exploring the unexplored section of the graph as much as it can be explored from the current node."